- implemented a vertex buffer for voxels. Faster but still not good enough. This needs more optimization...
git-svn-id: http://mancubus.net/svn/hosted/gzdoom/branches/voxels@1060 b0f79afe-0144-0410-b225-9a4edf0717df

+++ src/gl/models/gl_models.h
@@ -241,7 +241,7 @@ class FVoxelVertexBuffer;
 struct FVoxelVertex
 {
 	float x,y,z;
-	BYTE u,v;
+	float u,v;
 };
 
 struct FVoxelVertexHash
